+++ examples/test_linux.c
@@ -7,27 +7,27 @@ MultiTimer timer1;
 MultiTimer timer2;
 MultiTimer timer3;
 
-uint64_t PlatformTicksGetFunc(void)
+uint32_t PlatformTicksGetFunc(void)
 {
     struct timespec current_time;
     clock_gettime(CLOCK_MONOTONIC, &current_time);
-    return (uint64_t)((current_time.tv_sec * 1000) + (current_time.tv_nsec / 1000000));
+    return (uint32_t)((current_time.tv_sec * 1000) + (current_time.tv_nsec / 1000000));
 }
 
 void exampleTimer1Callback(MultiTimer* timer, void *userData)
 {
-    printf("[T:%012ld] Timer:%p callback-> %s.\r\n", PlatformTicksGetFunc(), timer, (char*)userData);
+    printf("[T:%010d] Timer:%p callback-> %s.\r\n", PlatformTicksGetFunc(), timer, (char*)userData);
     MultiTimerStart(timer, 1000, exampleTimer1Callback, userData);
 }
 
 void exampleTimer2Callback(MultiTimer* timer, void *userData)
 {
-    printf("[T:%012ld] Timer:%p callback-> %s.\r\n", PlatformTicksGetFunc(), timer, (char*)userData);
+    printf("[T:%010d] Timer:%p callback-> %s.\r\n", PlatformTicksGetFunc(), timer, (char*)userData);
 }
 
 void exampleTimer3Callback(MultiTimer* timer, void *userData)
 {
-    printf("[T:%012ld] Timer:%p callback-> %s.\r\n", PlatformTicksGetFunc(), timer, (char*)userData);
+    printf("[T:%010d] Timer:%p callback-> %s.\r\n", PlatformTicksGetFunc(), timer, (char*)userData);
     MultiTimerStart(timer, 4567, exampleTimer3Callback, userData);
 }
